+++ service/controller/resource/prometheus/resource.go
@@ -204,10 +204,9 @@ func toPrometheus(v interface{}, config Config) (metav1.Object, error) {
 			Storage: &storage,
 			Affinity: &v1.Affinity{
 				NodeAffinity: &v1.NodeAffinity{
-					PreferredDuringSchedulingIgnoredDuringExecution: []v1.PreferredSchedulingTerm{
-						v1.PreferredSchedulingTerm{
-							Weight: 60,
-							Preference: v1.NodeSelectorTerm{
+					RequiredDuringSchedulingIgnoredDuringExecution: &v1.NodeSelector{
+						NodeSelectorTerms: []v1.NodeSelectorTerm{
+							v1.NodeSelectorTerm{
 								MatchExpressions: []v1.NodeSelectorRequirement{
 									v1.NodeSelectorRequirement{
 										Key:      "role",