+++ service/controller/resource/resource.go
@@ -8,6 +8,7 @@ import (
 	"github.com/giantswarm/operatorkit/v2/pkg/resource/wrapper/metricsresource"
 	"github.com/giantswarm/operatorkit/v2/pkg/resource/wrapper/retryresource"
 	promclient "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned"
+	vpa_clientset "k8s.io/autoscaler/vertical-pod-autoscaler/pkg/client/clientset/versioned"
 
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alert"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/alertmanagerconfig"
@@ -21,6 +22,7 @@ import (
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/remotewriteconfig"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/scrapeconfigs"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/servicemonitor"
+	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/verticalpodautoscaler"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/volumeresizehack"
 	"github.com/giantswarm/prometheus-meta-operator/service/controller/resource/wrapper/monitoringdisabledresource"
 	"github.com/giantswarm/prometheus-meta-operator/service/key"
@@ -49,6 +51,7 @@ type Config struct {
 	K8sClient               k8sclient.Interface
 	Logger                  micrologger.Logger
 	PrometheusClient        promclient.Interface
+	VpaClient               vpa_clientset.Interface
 }
 
 func New(config Config) ([]resource.Interface, error) {
@@ -148,6 +151,19 @@ func New(config Config) ([]resource.Interface, error) {
 		}
 
 		prometheusResource, err = prometheus.New(c)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var verticalPodAutoScalerResource resource.Interface
+	{
+		c := verticalpodautoscaler.Config{
+			Logger:    config.Logger,
+			VpaClient: config.VpaClient,
+		}
+
+		verticalPodAutoScalerResource, err = verticalpodautoscaler.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -283,6 +299,7 @@ func New(config Config) ([]resource.Interface, error) {
 		scrapeConfigResource,
 		remoteWriteConfigResource,
 		prometheusResource,
+		verticalPodAutoScalerResource,
 		volumeResizeHack,
 		ingressResource,
 		promxyResource,